move widget store and client to their own modules
